requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0